Add files via upload
# Car Price Predictor

Project link: https://car-price-price.herokuapp.com

1) How to use?
Clone the repository
Install the required packages in "requirements.txt" file.
Some packages are:

numpy
pandas
scikit-learn
Run the "application.py" file And you are good to go.

2) What this project does?
This project takes the parameters of an used car like: Company name, Model name, Year of Purchase, Fuel Type and Number of Kilometers it has been driven.
It then predicts the possible price of the car. For example, the image below shows the predicted price of our Hyundai Grand i10.

3) How this project does?
First of all the data was scraped from Quikr.com (https://quikr.com) 
Link for data: 

The data was cleaned (it was super unclean :( ) and analysed.

Then a Linear Regression model was built on top of it which had 0.89 R2_score.

Link for notebook: 

This project was given the form of an website built on Flask where we used the Linear Regression model to perform predictions.